add EditText extension functions to own file
