+++ app/src/main/java/org/dhis2/Bindings/Extensions.kt
@@ -1,7 +1,5 @@
 package org.dhis2.Bindings
 
-import android.view.MotionEvent
-import android.widget.EditText
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.MutableLiveData
@@ -15,21 +13,6 @@ import org.hisp.dhis.android.core.trackedentity.TrackedEntityInstance
  * QUADRAM. Created by ppajuelo on 11/04/2019.
  */
 fun <T : Any?> MutableLiveData<T>.default(initialValue: T) = this.apply { setValue(initialValue) }
-
-fun EditText.onRightDrawableClicked(onClicked: (view: EditText) -> Unit) {
-    this.setOnTouchListener { v, event ->
-        var hasConsumed = false
-        if (v is EditText) {
-            if (event.x >= v.width - v.totalPaddingRight) {
-                if (event.action == MotionEvent.ACTION_UP) {
-                    onClicked(this)
-                }
-                hasConsumed = true
-            }
-        }
-        hasConsumed
-    }
-}
 
 fun List<Event>.toUids(): List<String> {
     return map { it.uid() }
